+++ app/src/main/java/com/example/moneyeverydayxml/domain/InteractorInterface.kt
@@ -1,7 +1,9 @@
 package com.example.moneyeverydayxml.domain
 
+import java.math.BigDecimal
+
 interface InteractorInterface {
-    fun saveData(amount: String, date: String, summary: Int)
+    fun saveData(amount: String, date: String, summary: BigDecimal)
     fun saveClearDate(clearDate: Long)
     fun loadData()
     fun getClearDate(): Long
